Removed outdated iteration of bytecode. Bytecode will return at a later date
