+++ src/umpteen/error.rs
@@ -1,11 +1,11 @@
 mod compiler_error;
 mod memory_error;
 mod parse_error;
-mod runtime_error;
+mod interpret_error;
 mod umpteen_error;
 
 pub use compiler_error::*;
 pub use memory_error::*;
 pub use parse_error::*;
-pub use runtime_error::*;
+pub use interpret_error::*;
 pub use umpteen_error::*;
