--- src/umpteen/repr/bytecode.rs
@@ -1,2 +1,0 @@
-pub mod chunk;
-pub mod instruction;
